User movement and position calculation

+++ Backend/Sources/App/Controllers/TunnelGenerator.swift
@@ -17,8 +17,8 @@ func generateTunnels() -> [Tunnel] {
         // FIXME: This is not optimal, because they will cluster in the middle
         let padding = Constants.FIELD_LENGTH / 10
         let virtualCenter = Position(
-            x: Int.random(in: padding ... (Constants.FIELD_LENGTH - padding)),
-            y: Int.random(in: padding ... (Constants.FIELD_LENGTH - padding))
+            x: Double(Int.random(in: padding ... (Constants.FIELD_LENGTH - padding))),
+            y: Double(Int.random(in: padding ... (Constants.FIELD_LENGTH - padding)))
         )
 
         let exits = (2 ... numExits).map { _ in
@@ -29,7 +29,7 @@ func generateTunnels() -> [Tunnel] {
             repeat {
                 position = Position(position: virtualCenter)
                 position.translate(
-                    r: Int.random(in: (Constants.EXIT_SIZE * 2) ... (Constants.EXITS_MAX_DISTANCE / 2)),
+                    r: Double.random(in: (Constants.EXIT_SIZE * 2) ... (Constants.EXITS_MAX_DISTANCE / 2)),
                     phi: Float64.random(in: 0 ... 2) * Float64.pi
                 )
             } while exitPositions.contains { ep in
